requirements.txt
Brotli==1.0.9
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.5.0
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
redis==3.5.3
python-dotenv==0.14.0